Add files via upload
In order to visit Acadia (Maine, USA), we want to find out the perfect time to plan a trip when is not crowded and flowers are blooming. We will make to plots for flights and blooming flowers and compare them at the end.  we apply basic skills of visualizing a distribution of quantitative data.